Downloads/idea/jtbd-agent/pyproject.toml
[project]
name = "jtbd_agent_dspy"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "prefect>=3.0.0",
  "requests>=2.32",
  "dspy-ai>=2.5.12",
  "fastapi>=0.111",
  "uvicorn>=0.30"
]

[tool.setuptools.packages.find]
where = ["."]
